Restore mt_evtEffect census source

+++ src/frTagTable.c
@@ -606,17 +606,17 @@ u64 FUN_003aed10(u64 param_1,int param_2)
 {
   int offset;
   int base;
-  u32 low;
-  u32 high;
+  u8 low;
+  u8 high;
   short sVar2;
   u32 uVar3;
-  u32 key;
+  int key;
   u8 *pcVar4;
 
   offset = *(volatile int *)((int)param_2 + 0x18);
   base = *(volatile int *)((int)param_2 + 0x10);
   pcVar4 = (u8 *)(base + offset);
-  low = (u8)(pcVar4[0] - 1);
+  low = pcVar4[0] - 1;
   high = pcVar4[1];
   if (high == 0xff) {
     high = 0;
